script: Split debian package installation out into functions
Previously this was rolled into the container build steps.

Splitting it out makes the required package lists clearer
to read, and simplifies the container build steps.

The package installation functions could also now be used
to install the required packages on the host system if the
user is doing it natively rather than with containers.

They would need to be on a Debian-based system and run with
sudo though.