script: Add ability to generate project Makefile
The generated Makefile uses SGDK Helper. The advantage of
using the Makefile as a wrapper is that it provides command
line tab completion for the targets.

We can also list multiple targets. For example:

    make clean rom run

Note: 'rom' is the default target, so when no targets are
specified, it will build the project ROM.